+++ packages/pangea-sdk/PangeaCyber.Net/AuthN/Requests/UserInviteListRequest.cs
@@ -1,5 +1,6 @@
 using Newtonsoft.Json;
 using PangeaCyber.Net.AuthN.Models;
+using PangeaCyber.Net.Filters;
 
 namespace PangeaCyber.Net.AuthN.Requests
 {
@@ -8,7 +9,7 @@ namespace PangeaCyber.Net.AuthN.Requests
     {
         ///
         [JsonProperty("filter")]
-        public Filter? Filter { get; private set; }
+        public FilterUserInviteList? Filter { get; private set; }
 
         ///
         [JsonProperty("last")]
@@ -28,18 +29,18 @@ namespace PangeaCyber.Net.AuthN.Requests
 
         private UserInviteListRequest(Builder builder)
         {
-            this.Filter = builder.Filter;
-            this.Last = builder.Last;
-            this.Order = builder.Order;
-            this.OrderBy = builder.OrderBy;
-            this.Size = builder.Size;
+            Filter = builder.Filter;
+            Last = builder.Last;
+            Order = builder.Order;
+            OrderBy = builder.OrderBy;
+            Size = builder.Size;
         }
 
         ///
         public class Builder
         {
             ///
-            public Filter? Filter { get; private set; }
+            public FilterUserInviteList? Filter { get; private set; }
             ///
             public string? Last { get; private set; }
             ///
@@ -52,38 +53,53 @@ namespace PangeaCyber.Net.AuthN.Requests
             ///
             public Builder() { }
 
-            ///
+            /// @deprecated use WithFilter(FilterUserInviteList filter) instead
             public Builder WithFilter(Filter filter)
             {
-                this.Filter = filter;
+                Filter = new FilterUserInviteList();
+                foreach (var kvp in filter)
+                {
+                    Filter.Add(kvp.Key, kvp.Value);
+                }
+                return this;
+            }
+
+            ///
+            public Builder WithFilter(FilterUserInviteList filter)
+            {
+                Filter = new FilterUserInviteList();
+                foreach (var kvp in filter)
+                {
+                    Filter.Add(kvp.Key, kvp.Value);
+                }
                 return this;
             }
 
             ///
             public Builder WithLast(string last)
             {
-                this.Last = last;
+                Last = last;
                 return this;
             }
 
             ///
             public Builder WithOrder(ItemOrder order)
             {
-                this.Order = order;
+                Order = order;
                 return this;
             }
 
             ///
             public Builder WithOrderBy(UserInviteListOrderBy orderBy)
             {
-                this.OrderBy = orderBy;
+                OrderBy = orderBy;
                 return this;
             }
 
             ///
             public Builder WithSize(int? size)
             {
-                this.Size = size;
+                Size = size;
                 return this;
             }