+++ packages/pangea-sdk/PangeaCyber.Net/BaseClient.cs
@@ -19,7 +19,7 @@ namespace PangeaCyber.Net
         private readonly string serviceName;
 
         ///
-        private readonly bool SupportMultiConfig;
+        public string ConfigID { get; protected set; } = "";
 
         ///
         protected readonly HttpClient HttpClient;
@@ -54,13 +54,12 @@ namespace PangeaCyber.Net
         }
 
         ///
-        protected BaseClient(ClientBuilder builder, string serviceName, bool SupportMultiConfig)
+        protected BaseClient(ClientBuilder builder, string serviceName)
         {
             this.config = builder.config;
             // Set default logger
             this.logger = builder.logger ?? this.GetDefaultLogger();
             this.serviceName = serviceName;
-            this.SupportMultiConfig = SupportMultiConfig;
             this.HttpClient = new HttpClient();
             this.HttpClient.BaseAddress = config.GetServiceUrl(serviceName, String.Empty);
 
@@ -86,35 +85,11 @@ namespace PangeaCyber.Net
             return logger;
         }
 
-        ///
-        public async Task<Response<TResult>> DoPost<TResult>(string path, BaseRequest request)
+        private async Task<HttpResponseMessage> DoPost(string path, HttpContent content)
         {
-            if (this.SupportMultiConfig && this.config.ConfigID != default && request.ConfigID == default)
-            {
-                request.ConfigID = this.config.ConfigID;
-            }
-
-            string requestStr;
-            StringContent requestJson;
             try
             {
-                var jsonSettings = new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore, DateParseHandling = DateParseHandling.None, DateFormatString = "yyyy'-'MM'-'dd'T'HH':'mm':'ss.fffffffK" };
-                requestStr = JsonConvert.SerializeObject(request, Formatting.Indented, jsonSettings);
-                requestJson = new StringContent(requestStr, Encoding.UTF8, "application/json");
-            }
-            catch (Exception e)
-            {
-                throw new PangeaException("Failed to write request", e);
-            }
-
-            this.logger.Debug(
-                $"{{\"service\": \"{serviceName}\", \"action\": \"post\", \"path\": \"{path}\", \"request\": {requestStr}}}"
-            );
-
-            HttpResponseMessage res = default!;
-            try
-            {
-                res = await this.HttpClient.PostAsync(path, requestJson);
+                return await this.HttpClient.PostAsync(path, content);
             }
             catch (Exception e)
             {
@@ -123,6 +98,50 @@ namespace PangeaCyber.Net
                 );
                 throw new PangeaException("Failed to send request", e);
             }
+        }
+
+        private string SerializeRequest(BaseRequest request)
+        {
+            try
+            {
+                var jsonSettings = new JsonSerializerSettings { NullValueHandling = NullValueHandling.Ignore, DateParseHandling = DateParseHandling.None, DateFormatString = "yyyy'-'MM'-'dd'T'HH':'mm':'ss.fffffffK" };
+                return JsonConvert.SerializeObject(request, Formatting.Indented, jsonSettings);
+            }
+            catch (Exception e)
+            {
+                throw new PangeaException("Failed to write request", e);
+            }
+        }
+
+        ///
+        public async Task<Response<TResult>> DoPost<TResult>(string path, BaseRequest request, FileStream? fileStream = null)
+        {
+            if (!string.IsNullOrEmpty(ConfigID) && string.IsNullOrEmpty(request.ConfigID))
+            {
+                request.ConfigID = ConfigID;
+            }
+
+            string requestStr = SerializeRequest(request);
+            this.logger.Debug(
+                $"{{\"service\": \"{serviceName}\", \"action\": \"post\", \"path\": \"{path}\", \"request\": {requestStr}}}"
+            );
+
+            HttpResponseMessage res = default!;
+
+            if (fileStream == null)
+            {
+                res = await DoPost(path, new StringContent(requestStr, Encoding.UTF8, "application/json"));
+            }
+            else
+            {
+                using (var formData = new MultipartFormDataContent())
+                {
+                    formData.Add(new StringContent(requestStr, null, "application/json"), "request");
+                    var fileContent = new StreamContent(fileStream);
+                    formData.Add(fileContent, "upload", "file.exe");
+                    res = await DoPost(path, formData);
+                }
+            }
 
             if (res.StatusCode == System.Net.HttpStatusCode.Accepted && this.config.QueuedRetryEnabled)
             {
@@ -130,6 +149,25 @@ namespace PangeaCyber.Net
             }
 
             return await CheckResponse<TResult>(res);
+        }
+
+        private async Task<Response<TResult>> DoPollResult<TResult>(string requestID)
+        {
+            string path = PollResultPath(requestID);
+            HttpResponseMessage res = await DoGet(path);
+            return await CheckResponse<TResult>(res);
+        }
+
+        ///
+        public async Task<Response<TResult>> PollResult<TResult>(string requestID)
+        {
+            return await DoPollResult<TResult>(requestID);
+        }
+
+        ///
+        public async Task<Response<Dictionary<string, object>>> PollResult(string requestID)
+        {
+            return await DoPollResult<Dictionary<string, object>>(requestID);
         }
 
         ///
@@ -339,6 +377,10 @@ namespace PangeaCyber.Net
                 else if (header.Status.Equals(ResponseStatus.IPNotFound.ToString()))
                 {
                     throw new EmbargoIPNotFoundException(summary, response);
+                }
+                else if (header.Status.Equals(ResponseStatus.Accepted.ToString()))
+                {
+                    throw new AcceptedRequestException($"Summary: \"{summary}\". request_id: \"{response.RequestId}\".", response);
                 }
             }
             throw new PangeaAPIException(String.Format("{0}: {1}", status, summary), response);