+++ packages/pangea-sdk/PangeaCyber.Net/Filters/FilterBase.cs
@@ -1,0 +1,19 @@
+namespace PangeaCyber.Net.Filters
+{
+    ///
+    public class FilterBase
+    {
+        ///
+        protected string Name;
+
+        ///
+        protected Filter Map;
+
+        ///
+        public FilterBase(string name, Filter map)
+        {
+            Name = name;
+            Map = map;
+        }
+    }
+}