+++ packages/pangea-sdk/PangeaCyber.Net/Redact/Requests/RedactTextRequest.cs
@@ -21,6 +21,10 @@ namespace PangeaCyber.Net.Redact
         public string[]? Rules { get; private set; }
 
         ///
+        [JsonProperty("rulesets")]
+        public string[]? Rulesets { get; private set; }
+
+        ///
         [JsonProperty("return_result")]
         public bool? ReturnResult { get; private set; }
 
@@ -30,6 +34,7 @@ namespace PangeaCyber.Net.Redact
             this.Text = builder.Text;
             this.Debug = builder.Debug;
             this.Rules = builder.Rules;
+            this.Rulesets = builder.Rulesets;
             this.ReturnResult = builder.ReturnResult;
         }
 
@@ -47,6 +52,9 @@ namespace PangeaCyber.Net.Redact
 
             ///
             public string[]? Rules { get; private set; }
+
+            ///
+            public string[]? Rulesets { get; private set; }
 
             ///
             public bool? ReturnResult { get; private set; }
@@ -68,6 +76,13 @@ namespace PangeaCyber.Net.Redact
             public Builder WithRules(string[] rules)
             {
                 this.Rules = rules;
+                return this;
+            }
+
+            ///
+            public Builder WithRulesets(string[] rulesets)
+            {
+                this.Rulesets = rulesets;
                 return this;
             }