+++ packages/pangea-sdk/PangeaCyber.Net/Utils/Convertions.cs
@@ -18,12 +18,24 @@ namespace PangeaCyber.Net
         }
 
         ///
-        public static string Base64ToString(string input){
+        public static string Base64ToString(string input)
+        {
             // Convert Base64 string to byte array
             byte[] bytes = Convert.FromBase64String(input);
 
             // Decode byte array to string
             return Encoding.UTF8.GetString(bytes);
+        }
+
+        ///
+        public static string Bytes2Hex(byte[] bytes)
+        {
+            StringBuilder builder = new StringBuilder();
+            foreach (byte b in bytes)
+            {
+                builder.Append(b.ToString("x2")); // Convert each byte to its hexadecimal representation
+            }
+            return builder.ToString();
         }
 
         ///
@@ -33,20 +45,62 @@ namespace PangeaCyber.Net
             {
                 byte[] inputBytes = Encoding.UTF8.GetBytes(input);
                 byte[] hashBytes = sha256.ComputeHash(inputBytes);
-                StringBuilder builder = new StringBuilder();
-
-                foreach (byte b in hashBytes)
-                {
-                    builder.Append(b.ToString("x2")); // Convert each byte to its hexadecimal representation
-                }
-
-                return builder.ToString();
+                return Bytes2Hex(hashBytes);
             }
         }
 
         ///
-        public static string GetHashPrefix(string input, int len = 5){
-            if(string.IsNullOrEmpty(input))
+        public static string GetSHA1Hash(string input)
+        {
+            using (SHA1 sha1 = SHA1.Create())
+            {
+                byte[] inputBytes = Encoding.UTF8.GetBytes(input);
+                byte[] hashBytes = sha1.ComputeHash(inputBytes);
+                return Bytes2Hex(hashBytes);
+            }
+        }
+
+        ///
+        public static string GetSHA512Hash(string input)
+        {
+            using (SHA512 sha1 = SHA512.Create())
+            {
+                byte[] inputBytes = Encoding.UTF8.GetBytes(input);
+                byte[] hashBytes = sha1.ComputeHash(inputBytes);
+                return Bytes2Hex(hashBytes);
+            }
+        }
+
+        ///
+        public static string GetSHA256HashFromFilepath(string filepath)
+        {
+            using (FileStream stream = File.OpenRead(filepath))
+            {
+                using (SHA256 sha256 = SHA256.Create())
+                {
+                    byte[] hashBytes = sha256.ComputeHash(stream);
+                    return Bytes2Hex(hashBytes);
+                }
+            }
+        }
+
+        ///
+        public static string GetSHA1HashFromFilepath(string filepath)
+        {
+            using (FileStream stream = File.OpenRead(filepath))
+            {
+                using (SHA1 sha1 = SHA1.Create())
+                {
+                    byte[] hashBytes = sha1.ComputeHash(stream);
+                    return Bytes2Hex(hashBytes);
+                }
+            }
+        }
+
+        ///
+        public static string GetHashPrefix(string input, int len = 5)
+        {
+            if (string.IsNullOrEmpty(input))
             {
                 return input;
             }