+++ packages/pangea-sdk/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2023-09-05
+
+# Added
+
+- Redact rulesets field support 
+- FileScan service support
+
+
 ## [2.2.0] - 2023-07-28 
 
 # Added
@@ -46,7 +54,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Redact requests builder rename to be called just "Builder" 
 
 
-[unreleased]: https://github.com/pangeacyber/pangea-dotnet/compare/v2.2.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-dotnet/compare/v2.3.0...main
+[2.3.0]: https://github.com/pangeacyber/pangea-dotnet/compare/v2.2.0...v2.3.0
 [2.2.0]: https://github.com/pangeacyber/pangea-dotnet/compare/v2.1.0...v2.2.0
 [2.1.0]: https://github.com/pangeacyber/pangea-dotnet/compare/v2.0.0...v2.1.0
 [2.0.0]: https://github.com/pangeacyber/pangea-dotnet/releases/tag/v2.0.0